exception added

+++ src/article/article.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateArticleDto } from './dto/create-article.dto';
 import { UpdateArticleDto } from './dto/update-article.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -14,15 +14,27 @@ export class ArticleService {
     return this.prisma.article.findMany({ where: { published: true } });
   }
 
-  findOne(id: number) {
-    return this.prisma.article.findUnique({ where: { id } });
+  async findOne(id: number) {
+    const article = await this.prisma.article.findUnique({ where: { id } });
+    if (!article) {
+      throw new NotFoundException(`Article with ${id} does not exist.`);
+    }
+    return article;
   }
 
-  update(id: number, updateArticleDto: UpdateArticleDto) {
+  async update(id: number, updateArticleDto: UpdateArticleDto) {
+    const article = await this.prisma.article.findUnique({ where: { id } });
+    if (!article) {
+      throw new NotFoundException(`Article with ${id} does not exist.`);
+    }
     return this.prisma.article.update({ where: { id }, data: updateArticleDto });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const article = await this.prisma.article.findUnique({ where: { id } });
+    if (!article) {
+      throw new NotFoundException(`Article with ${id} does not exist.`);
+    }
     return this.prisma.article.delete({ where: { id } });
   }
 }
